% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/culture_plates.R
\name{culture_plate}
\alias{culture_plate}
\title{Draw plate and color by condition}
\usage{
culture_plate(
  data,
  condition_columns,
  color = default_color,
  plate_type = "96 wells",
  ref = default_dimension
)
}
\arguments{
\item{data}{data frame of plate}

\item{condition_columns}{columns name for condition}

\item{color}{colors for condtion to use}

\item{plate_type}{Plate type e.g. 6 wells}

\item{ref}{dimension for plates}
}
\value{
ggplot of plate
}
\description{
Draw plate and color by condition
}
\examples{
culture_plate_empty("6 wells")
}
