% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/culture_plates.R
\name{culture_plate_empty}
\alias{culture_plate_empty}
\title{Draw empty plate}
\usage{
culture_plate_empty(plate_type, ref = default_dimension)
}
\arguments{
\item{plate_type}{Plate type e.g. 6 wells}

\item{ref}{dimension for plates}
}
\value{
ggplot of plate
}
\description{
Draw empty plate
}
\examples{
culture_plate_empty("6 wells")
}
